+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Version jump from 0.3->0.7 is to align with other RAPIDS projects.
 - PR #176 Added match_strings to nvstrings
 - PR #172 Added to/from boolean conversion methods
 - PR #171 Added conversion to/from subset of ISO8601 format
+- PR #230 Add ngrams function to nvtext module
 
 ## Improvements